requirements.txt
asgiref==3.5.0
Babel==2.10.1
CacheControl==0.12.11
certifi==2021.10.8
charset-normalizer==2.0.12
deprecation==2.1.0
dj-database-url==0.5.0
Django==3.2.13
django-cors-headers==3.11.0
django-filter==21.1
django-heroku==0.3.1
django-money==2.1.1
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
jsonpickle==1.5.2
Markdown==3.3.6
msgpack==1.0.3
packaging==21.3
psycopg2==2.9.3
py-moneyed==1.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
squareup==19.0.0.20220512
typing-extensions==4.2.0
urllib3==1.26.9
whitenoise==6.0.0
zipp==3.8.0
